requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord-protos==0.0.2
discord.py-self==2.0.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.5.0
protobuf==4.23.2
requests==2.31.0
urllib3==2.0.2
yarl==1.9.2
